+++ Sources/MercadoPagoCheckout/Views/CardFormScreen.swift
@@ -22,6 +22,10 @@ struct CardFormScreen: View {
     // Formatters and Validators
     private let cardNumberFormatter = CardNumberFormatter()
     private let cardNumberValidator = CardNumberValidator()
+    private let expirationDateFormatter = ExpirationDateFormatter()
+    private let expirationDateValidator = ExpirationDateValidator()
+    private let securityCodeFormatter = SecurityCodeFormatter()
+    private let securityCodeValidator = SecurityCodeValidator()
     
     //  Document Field
     @State private var selectTypeDocument: IdentificationType = .init(name: "CPF")
@@ -63,12 +67,17 @@ struct CardFormScreen: View {
                         placeholder: MPStrings.CardForm.CardHolder.placeholder
                     )
                     
-                    HStack(spacing: theme.spacings.xl) {
+                    HStack(
+                        alignment: .top,
+                        spacing: theme.spacings.xl
+                    ) {
                         MPTextField(
                             text: $expirationDate,
                             label: MPStrings.CardForm.Expiration.label,
                             placeholder: MPStrings.CardForm.Expiration.placeholder,
-                            keyboard: .numberPad
+                            keyboard: .numberPad,
+                            formatter: expirationDateFormatter,
+                            validator: expirationDateValidator
                         )
                         
                         MPTextField(
@@ -76,12 +85,13 @@ struct CardFormScreen: View {
                             label: MPStrings.CardForm.CVV.label,
                             placeholder: MPStrings.CardForm.CVV.placeholderDefault,
                             keyboard: .numberPad,
+                            formatter: securityCodeFormatter,
+                            validator: securityCodeValidator,
                             suffix: {
                                 Image(systemName: "questionmark.circle")
                                     .renderingMode(.template)
                                     .foregroundColor(theme.colors.accent)
                                     .padding(.horizontal, theme.spacings.s)
-
                             }
                         )
                     }
